+++ workspace/plugins/jupyter-client/src/index.ts
@@ -117,6 +117,7 @@ import type {
     IClickEditorTitleIconEvent,
     IDestroyProtyleEvent,
     ILoadedProtyleStaticEvent,
+    ISwitchProtyleEvent,
 } from "@workspace/types/siyuan/events";
 import type { THandlersWrapper } from "@workspace/utils/worker/bridge";
 import type { WorkerHandlers } from "./workers/jupyter";
@@ -521,7 +522,8 @@ export default class JupyterClientPlugin extends siyuan.Plugin {
                 this.eventBus.on("click-editortitleicon", this.blockMenuEventListener);
                 this.eventBus.on("click-blockicon", this.blockMenuEventListener);
                 this.eventBus.on("click-editorcontent", this.clickEditorContentEventListener);
-                this.eventBus.on("loaded-protyle-static", this.loadedProtyleStaticEventListener);
+                this.eventBus.on("loaded-protyle-static", this.loadedProtyleEventListener);
+                this.eventBus.on("switch-protyle", this.loadedProtyleEventListener);
                 this.eventBus.off("destroy-protyle", this.destroyProtyleEventListener);
             });
     }
@@ -567,7 +569,8 @@ export default class JupyterClientPlugin extends siyuan.Plugin {
         this.eventBus.off("click-editortitleicon", this.blockMenuEventListener);
         this.eventBus.off("click-blockicon", this.blockMenuEventListener);
         this.eventBus.off("click-editorcontent", this.clickEditorContentEventListener);
-        this.eventBus.off("loaded-protyle-static", this.loadedProtyleStaticEventListener);
+        this.eventBus.off("loaded-protyle-static", this.loadedProtyleEventListener);
+        this.eventBus.off("switch-protyle", this.loadedProtyleEventListener);
         this.eventBus.off("destroy-protyle", this.destroyProtyleEventListener);
 
         for (const objectURL of this.kernelName2logoObjectURL.values()) {
@@ -2166,7 +2169,7 @@ export default class JupyterClientPlugin extends siyuan.Plugin {
     }
 
     /* 编辑器加载事件监听器 */
-    protected readonly loadedProtyleStaticEventListener = async (e: ILoadedProtyleStaticEvent) => {
+    protected readonly loadedProtyleEventListener = async (e: ILoadedProtyleStaticEvent | ISwitchProtyleEvent) => {
         // this.logger.debug(e);
         const protyle = e.detail.protyle;